Add HighCardinalityTagsDetector (#3047)
Introduces the concept of a detector for high cardinality tags. High cardinality tags are generally a problem for metrics, and except in special circumstances, it is expected that tags on metrics will be low cardinality or otherwise the cardinality bounded. It is an instrumentation error to make high cardinality tags. A library you are using may have Micrometer instrumentation and be contributing high cardinality tags without you explicitly instrumenting anything. This detector can help identify the meter that has high cardinality tags if you have noticed high memory usage or other signs of high cardinality tags.