Avoid classloading LatencyUtils when not configured (#3262)
The default configuration uses a NoPauseDetector, which need not use any LatencyUtils classes. This avoids classloading LatencyUtils classes, which makes it possible for end users to exclude the LatencyUtils dependency and not have any issues at runtime.

The Prometheus sample should work without LatencyUtils or HdrHistogram since configuration that would use either is not present in the sample. This sample can be run to demonstrate it being possible to exclude these dependencies.