Merge service files in shadowJar
The service loader won't work with the package names before relocation unless a user happens to have those on the classpath, in which case the service descriptor file will clash with what is already on the user's classpath. We should use our relocated shaded package and combine service descriptor files if multiple exist.

Resolves gh-3268