Http client instrumentation TCK (#3258)
Introduce test suite for HTTP client instrumentations to check that all implementations have the naming and tags expected. This can also help instrumentors ensure backwards compatibility of the HTTP client instrumentation across changes to it. Introduced as incubating until we get feedback from instrumentation outside of Micrometer on the usability and usefulness of the TCK in this form.