WIP on Receiver and SenderContext
With this change we are spiking complete removal of any http / messaging / rpc abstractions and leaving a ReceiverContext and RequestReply or FireAndForget SenderContext. That way we are able to propagate any content over the wire and we are not forcing any default tagging mechanism (other than the one to be backward compatible).
It turned out that we're not properly supporting OkHttp instrumentation so I had to rewrite it to use the Interceptor approach not the EventListener one (which simplified a lot of things too).

Thing to consider: do we even need the transport.http package - should we remove it too? Those extend the Observation.KeyValuesConvention and could be provided by external libraries.