Added option to set a parent Observation on an existing Observation (#3291)
* Added option to set a parent Observation on an existing Observation

without this change in order to e.g. create spans we always need to put an observation in scope so that spans are put in thread local and a child span is being set. That doesn't make a lot of sense when we exactly know what the parent observation is.
with this change we're allowing to set a parent observation direcly on an existing observation. That way we're able to maintain a parent child relationship of observations and of e.g. spans

* Polish

* When using scope parent observation should be locally set only