Receiver and SenderContext (#3293)
With this change we are spiking complete removal of any http / messaging / rpc abstractions and leaving a  Receiver & Sender contexts (with a request-reply option). That way we are able to propagate any content over the wire and we are not forcing any default tagging mechanism (other than the one to be backward compatible).

It turned out that we're not properly supporting OkHttp instrumentation so I had to rewrite it to use the Interceptor approach not the EventListener one (which simplified a lot of things too).