Generify throwable in CheckedCallable/CheckedRunnable (#3326)
This allows usage that catches or throws a more specific type than `Throwable`. This is a bit more verbose when declaring a variable of type `CheckedCallable` or `CheckedRunnable`, but I expect most usage to be as a lambda or method reference, which avoids the need to do that.

Existing `CheckedCallable` types declared will fail to compile after this change, and they will need to add the type of the `Throwable` to the type declaration. This is effectively optional for `CheckedRunnable` since it can be used as its raw type.