Declare HdrHistogram as a runtime dependency (#3263)
HdrHistogram is not exposed in API users should be compiling against. It is an implementation detail of client-side percentiles. The scope of the dependency can be changed accordingly. If client-side percentiles are used, HdrHistogram is needed at runtime, as declared with this change. If client-side percentiles are not used, HdrHistogram can be excluded from the runtime classpath without issue.