Add option to disable step allignment of the report in PushRegistryConfig
The problem with the current implementation is that if we use a 60sec (1min) step by default, all the reporting will happen at the same time across all the application on the globe.
That cause huge issues for backends or intermediary services, such as Otel Collector which has to handle a large burst, and because of that a larger footprint is needed.

Signed-off-by: Bogdan Drutu <bogdandrutu@gmail.com>