Remove Flux dependency from statsd meters
The current implementation of the statsd registry is based on the Flux
reactor framework for handling events. This detail is leaked into the
implementation of the meters by passing a Flux.sink into the meters
constructor, making it hard to re-use the meter in any other fashion.
This changes swaps out the Flux.sink with a generic
java.util.function.Consumer, making the implementations more generically
useful.