JDK11 HttpClient instrumentation (#3358)
Adds instrumentation for the JDK HttpClient that was added in JDK 11. Since our modules are still compatible with Java 8, the instrumentation for the HttpClient is in a separate jdk11 source set in micrometer-core, which is packaged in the micrometer-core jar as a multi-release jar. This raises the required build JDK to 11. Accordingly, the CI build for JDK 8 has been removed.

Co-authored-by: Jonatan Ivanov <jonatan.ivanov@gmail.com>
Co-authored-by: Tommy Ludwig <8924140+shakuzen@users.noreply.github.com>