Remove conditional check for disabling japicmp in otlp
The version is not determined correctly in release builds from tags. It ends up being `1.10.0-SNAPSHOT` for instance, when doing a patch release of 1.9.x. We could probably rework the release process to avoid this issue, but this change is less risky given we are trying to do a release now.

Workaround due to problem described in gh-3408