Automated java module checking (#3411)
Adds a task called `testModules` that calls `java --list-modules` on each JAR file built to check that its modules can be listed. We have had issues in the past that went unnoticed, and this is one of the easiest ways to do a sanity check that a JAR will work on the module path. See, for example, gh-3398 and gh-3412, which would have been caught by this check.