TCK with observation
Needed to update JettyClientKeyValues to handle the case when the result was null, which will always happen when getting the low cardinality tags for the LongTaskTimer at the start of timing. For now, I've updated to return the full set of tags with an UNKNOWN value for things that can't be derived with a null result.