Rename DocumentedObservation to ObservationDocumentation
The previous naming implied that it was a type of `Observation` but it is rather the documentation for observations. The goal of this new naming is to make that more clear. It also improves consistency with the naming of most of the methods, though some inconsistency remains, which we can address in a follow-up change.