Take Supplier<Context> instead of Context (#3443)
* WIP Take Supplier<Context> instead of Context

This allows avoiding instantiation of a Context object in the case that the ObservationRegistry is a no-op. This does not overload the methods by having a variant that takes a Context and one that takes a Supplier because this becomes ambiguous for the compiler in some cases. A lambda can be used in cases where you have a Context object to pass. Alternatively, you can implement Supplier in your custom Context class to avoid needing a lambda.

* Use method reference instead of lambda returning null

* Polish

* Polish

* Make sure context supplier only called once

* Remove wildcard generics where it doesn't make sense

When not providing a context, one of type `Context` will be made. It doesn't make sense to take an `ObservationConvention` targeting a more specific type, then.

* JavaDocs updated with an explanation of no-op behavior