Allow documenting optional keys
There is no distinction between required and optional keys currently when using ObservationDocumentation. This introduces a new method `isRequired` that will default to true. Optional keys may be documented by overriding this method.

Future use of this can be in the docs-generator marking whether a key is required or not in the generated documentation.
Also, testing the instrumentation with our TCKs can check that required keys are always present, and that undocumented keys are not present but optional keys may or may not be present.