Bump to a medium+ resource class for builds (#3455)
Our builds are running at max CPU and RAM for most of the CI build on the default `medium` resource class for docker executors. We could probably get some improvement from more resources. This bumps up one level to `medium+` and makes the resource class choice explicit in the build, instead of the implicit default.