Test against ObservationDocumentation in instrumentation TCK (#3372)
Instrumentation being tested against the TCK can optionally provide the ObservationDocumentation that specifies the expected semantic naming to be used in the instrumentation. This ensures that the implementation matches the specified documentation, which may be used to generate written documentation.

* Introduce InstrumentationTimingVerificationTests

This adds a new abstract class in between `InstrumentationVerificationTests` and the `HttpClientTimingInstrumentationVerificationTests`/`HttpServerTimingInstrumentationVerificationTests`. The class `InstrumentationTimingVerificationTests` contains methods and logic specifically for timing instrumentation. The goal is to leave `InstrumentationVerificationTests` generic enough that it may be used for other kinds of instrumentation verification.