Flip default to align publishing with registry instantiation
We know the previous behavior causes issues with many instances running, which makes it not a good default. We will leave this configuration for now so that users have a way to opt out of the new behavior if it causes unforeseen issues for them.
Adds a warning to the JavaDoc about the side effect of aligning to the Epoch.