Jetty 11 module (#3382)
Introduces a new Jetty 11 instrumentation module. This is a separate module because Jetty 11 requires Java 11 and it has breaking changes from Jetty 9 that we cannot adapt to in the same module (such as the Jakarta namespace change). The implementation is copy and paste from the existing Jetty 9 instrumentation, adapting only to the API changes in Jetty 11.

Resolves gh-3234