Support "_index_template" for Elasticsearch 7.8+
Prior to this commit, the `ElasticMeterRegistry` would start exporting
metrics as documents after creating an index template to be used for the
metrics indices. The `"/_template"` would be used to send the mappings
configuration for those indices.

As of Elasticsearch 7.8, the new "composable index templates" supersede
the former, which is now deprecated. They are now more modular and can
reuse component templates. As of Elasticsearch 7.8, if a composable
index template and a legacy one exist under the same name, the new
variant will be used.

This change adds support for creating a composable index template for
metrics indices, if the feature is supported by the Elasticsearch
instance. The index template creation code has been extracted from
`ElasticMeterRegistry` in a `LegacyIndexTemplateCreator` strategy. The
new `DefaultIndexTemplateCreator` implementation uses not only a
different endpoint, `"/_index_template/"`, but also requires a different
document structure for describing the template. The meter registry will
now try to use the default strategy and then the legacy one as a
fallback.

This commit also updates the integration test to use the latest 8.6.x
version and the latest 7.7.x version (which did not support the new
feature) to ensure compatibility with older versions.

Closes gh-3320