Add BND plugin and OSGi export core package. (#3457)
Uses the BND plugin to enable the generation of OSGi metadata for all JARs (so turns them into OSGi bundles). Tests with OSGi Test to validate that the generated bundles resolve in an OSGi framework.
Note: only the core, JMX and Prometheus bundles are tested. Some bundles are likely to fail if they have dependencies that are also missing the OSGi metadata.

Resolves gh-3149