Avoid concurrent publish calls from publishSafely (#3714)
Guards calling `publish` in `publishSafely` with an `isPublishing` atomic flag. This will prevent the publish on close from happening if a scheduled publish is already in progress. It does not protect against calls to the `publish` method directly since that is implemented by children of PushMeterRegistry. This also exposes a protected `isPublishing` method that child classes can use if they want to do any conditional logic on this new flag. `StepMeterRegistry` will use this to avoid calling manualRollover on close if already publishing.

Resolves gh-3711

Co-authored-by: sonatype-lift[bot] <37194012+sonatype-lift[bot]@users.noreply.github.com>
Co-authored-by: Tommy Ludwig <8924140+shakuzen@users.noreply.github.com>