Do not rollover again after closingRollover
If a registry is closed near the end of a step interval, the publish on closing may take enough time that some meters are read after the step boundary is crossed. Without this fix, that results in those meters rolling over again, which we do not want in this case.

This renames manualRollover to closingRollover to signal this difference in behavior.

Resolves gh-3720