Allow resolving context-propagation milestone versions
The snapshot repo was removed in the `1.10.x` branch so snapshots would not be resolved. However, since `main` is trying to get the `1.1.+` version range of the context-propagation dependency, it could not resolve it. The error message was very misleading, but this resolves the issue by allowing milestone versions to be resolved.