Improve Otlp Delta Aggregation with support for max and Histogram (#3749)
Introduces a StepMax that rolls over with the step interval to match the OTLP specification for max in histograms. Updates the OTLP delta meters to use this. Also introduces a StepBucketHistogram that likewise aligns the rotation of the histogram buckets to the step interval, which is what the OTLP specification calls for. Some updates were made to abstract classes to enable the customization necessary for using these types.

Co-authored-by: Tommy Ludwig <8924140+shakuzen@users.noreply.github.com>