Add time units to OTel integration tests
The OTel collector had/have multiple bugs around this:
- The time unit is not converted to `seconds` (Prometheus' default)
- The time unit was not even visible in the name of the time series,
this violated the OTel specification

This means that if you sent 123ms to the OTLP collector,
on Prometheus side this turned into 123s
(the value is the same and not having the unit means `seconds`).
See: https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/18903
It seems units are still not converted but at least the unit is in the
name now (breaking change),
see: https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/20519

Because of this breaking change, our tests are also broken and we need to add the unit to our assertions.

Closes gh-3796