Pass toolchain Java version to osgi test
This was using the version of Java that Gradle itself was running on, which is often the same as the one used by the toolchain, but it is not always. Such a case was when we were trying JDK 21 early access builds, which Gradle itself would not run on at the time. The Java version passed to the osgi test.bndrun was not correct in this case, causing a build failure.
This passes the Java version configured with the toolchain.

Resolves gh-3869