Fix histogram bucket count with publishPercentileHistogram
There was a mismatch between the underlying buckets used in a histogram and the buckets that would be returned in a HistogramSnapshot. This resulted in the wrong counts due to a change in the way we calculate the bucket count, which assumes underlying buckets = buckets to return. This change makes the underlying buckets used the same as the buckets that will be returned, since there is no benefit to using more buckets than we will return.

A test has been added to the TCK to cover this case. It needs to be overly flexible since we don't know at the TCK level whether a MeterRegistry uses cumulative bucket counts or whether it supports aggregable percentile histogram buckets.