Remove unnecessary code
The check and adding of the percentile histogram buckets was redundant because this is already done in DistributionStatisticConfig#getHistogramBuckets with the proper check.

The tests in OtlpMeterRegistryTest were not failing before the change to main code and are duplicating the tests in the TCK (albeit a bit more specific on asserting).