Constant version of otel collector
We should have reproducible builds and maintenance branches frequently breaking on new versions of the otel collector creates a lot of work for us. This fixes the version and manages it in the build file. It's not the same as other dependencies are managed (no lock file), but it's maybe an improvement and keeps versions in build files rather than in a Java test class.
Once merged to `main` it can be set to `latest` until that becomes a maintenance branch.

This does add a little work to run the integration test from an IDE because the system property configuration needs to be added to your IDE run config.