Add ObservationExecChainHandler for HttpComponents 5.x
Prior to this commit, the instrumentation for HttpComponents 5.x would
rely on infrastructure that would not always be aware of runtime errors
or cancel signals; this could result in memory leaks or instrumentation
errors at runtime.

This commit introduces a new `ObservationExecChainHandler` which
implements both `ExecChainHandler` and `AsyncExecChainHandler`
contracts. This allows to instrument both "classic" and "async" clients
without the limitations underlined above.

This commit deprecates the former instrumentation as a result.

Fixes gh-3800