Introduce HttpJakartaServletRequestTags
Prior to this commit, `HttpRequestTags` would provide Tags for both
"javax.*" and "jakarta.*" variants of `HttpServletRequest`. While the
`HttpServletRequestTagsProvider` and
`HttpJakartaServletRequestTagsProvider` are separate interfaces, they
both use `HttpRequestTags` and import all types exposed on its methods.

It is strongly advised not to import both "javax.*" and "jakarta.*"
variants in a typical application and the current arrangement prevents
this best practice.

This commit deprecates Jakarta methods variants in `HttpRequestTags` and
moves them to `HttpJakartaServletRequestTags`. This allows a clean
dependency setup for Jakarta apps and a migration path for all in the
meantime.

Closes gh-3804