Convert to version catalog TOML file
Centralizes dependency version declaration to the version catalog toml file for type-safe reference of versions and to support automated dependency upgrades.
The goal here is to not change any dependencies from before, but move things to the toml file. We can follow-up with more refactors once we are confident these changes haven't inadvertently introduced unexpected behavior.