Escape/Unescape JMS String properties in contexts
Prior to this commit, both `JmsPublishObservationContext` and
`JmsProcessObservationContext` would set propagation headers as String
properties on the JMS `Message`.
This would cause issues with JMS implementations, as the JMS
specification only allows valid Java identifiers as keys. Typically, a
`X-B3-TraceId` header fails this requirement as it contains "-"
characters.

This commit ensures that the following characters are escaped when
publishing messages, and unescaped when receiving messages:

* "-" becomes "_HYPHEN_"
* "." becomes "_DOT_"

Closes gh-4202