Only serialize metadata when set on the meter
gh-4006 added support for metrics metadata (unit and description) in the
Dynatrace Exporter. The exporter always added the metadata even if unit
or description were not set. After these changes,
the exporter should not attempt to add metadata if no unit
or description was set on the metric.

See gh-4247
See gh-3979
See gh-4006

Co-authored-by: Jonatan Ivanov <jonatan.ivanov@gmail.com>