Reduce dependabot frequency to weekly
We are getting a lot of dependency upgrade pull requests. There is no benefit to users of any upgrades until we make a release. Our release schedule is monthly, so the previous setting of checking for updates daily is overly distracting. Monthly would be sufficient if it ran soon before our schedule releases, but since we cannot control the exact schedule of when dependabot runs, we will compromise with weekly updates.