Allow using TomcatMetrics without `Manager` on the compilation classpath
The two static factory methods allow users to create an instance of
`TomcatMetrics` without requiring them to have
`org.apache.catalina.Manager` on the compilation classpath e.g. as a
dependency with `provided` scope.

This is useful for for standalone Tomcat deployments where a war
artifact is being deployed. In most cases the application code will not
directly depend on Tomcat code but rather on javax/jakarkta APIs.