Wait for in-progress publish on close
When close is called, if a publish is not currently happening, the thread calling close will perform a publish and not return until it finishes. Before the changes in this commit, if a publish was in progress, close would not wait for it to finish, which was problematic in the case of close being called on application shutdown. This behavior was inconsistent in these two scenarios. This changes the behavior to wait for an in-progress publish to finish when close is called.