Switch dependabot to monthly
Reduce the deluge of pull requests from dependabot we need to deal with by reducing the frequency. Our scheduled releases are once a month, and we can manually trigger dependabot from https://github.com/micrometer-metrics/micrometer/network/updates prior to a release if we want to check for new versions.