Wait for in-progress publish before closing step rollover
If a scheduled publish is in progress when `close` is called, `StepMeterRegistry` was not performing the closing rollover and was not doing a subsequent publish of the final partial step.

Wait for the in-progress publish to finish before the closing rollover is called, which is followed by a final publish of the partial step.

Fixes gh-3846