Downgrade okhttp to non-alpha 4.12.0
We were on alpha versions as a result of using `latest.release` and the alphas being published to Maven central. When migrating to the TOML version catalog we kept the version the same as before. This attempts to move us off the alpha versions onto GA versions. We should not depend on any new API introduced in 5.0 alphas.