Downgrade okhttp to 4.11.0 upgrade Kotlin to 1.9 (#4597)
OkHttp 4.12.0 depends on a version of okio that ends up requiring Kotlin 1.9. We only have an optional dependency on OkHttp so we could upgrade to 4.12.0, but I would rather not compile against a newer version of OkHttp than our users can use if they are on our minimum supported version of Kotlin (1.7). For that reason, compiles against the latest version of OkHttp that is compatible with Kotlin 1.7 - OkHttp 4.11.0.
Uses Kotlin 1.9 to compile but remains compatible with Kotlin 1.7.