Polish gh-4753
Makes the Jetty server dependency optional. We plan to add Jetty client instrumentation and users may be use only one or the other, so micrometer-jetty12 should not pull in both dependencies. It will be up to users to separately have a dependency on Jetty server or client.

Updated copyright years and since tags.

Removed incubating annotations.