Count Generational ZGC cycles as concurrent time
This was missed when support for Generational ZGC was otherwise added in gh-4258. Non-generational ZGC uses the collector name `ZGC Cycles` for concurrent phases. Generational ZGC however uses `ZGC Minor Cycles` and `ZGC Major Cycles` which was not included in prior conditional for ZGC.

This conditionally enables a previously disabled test and makes sure it accounts for any GC that happens before metrics are recorded. It only works with a limited number of collectors (ZGC and Shenandoah) but it would have caught this issue and will ensure we don't introduce regressions.

Resolves gh-4764