Ignore class level AOP annotation if method level present (#3811)
Previously, the method would be instrumented twice if there was an annotation at the class and method level. This is generally not what users want, and as a matter of best practice, we do not want to instrument the same thing multiple times. With this change, the class level annotation will be ignored if there is a method level annotation, resulting in instrumenting only once with the configuration on the method level annotation.