Set description for CloudWatch modules
Instead of using the description set in the root build.gradle, sets a specific description for the CloudWatch modules. Calls out the deprecation of the v1 module in the description for better visibility to users. This description goes in the POM and is displayed in places like the web view for searching Maven Central.

Closes gh-4831